source/requirements-dev.txt
avro==1.11.1
aws-lambda-powertools==1.30.0
aws_cdk_lib==2.44.0
aws-cdk.aws-servicecatalogappregistry-alpha==2.44.0a0
aws_solutions_constructs.aws_eventbridge_sqs~=2.25.0
aws_solutions_constructs.aws_lambda_sns~=2.25.0
black
boto3
requests==2.28.1
requests-oauthlib
chalice
constructs>=10.0.0,<11.0.0
requests-mock==1.10.0
crhelper~=2.0.10
cronex==0.1.3.1
moto==4.0.5
path
pandas
parsedatetime==2.6
pytest
pytest-cov>=2.11.1
pytest-env>=0.6.2
pytest-mock>=3.5.1
pyyaml==6.0.0
responses~=0.17.0
-e cdk_solution_helper_py/helpers_cdk
-e cdk_solution_helper_py/helpers_common
-e infrastructure
-e aws_lambda
-e api